Fixed console spam with linking shaders, fixed NPE with ModelRendererWavefront.

+++ src/main/java/com/zeitheron/hammercore/client/model/mc/ModelRendererWavefront.java
@@ -163,7 +163,8 @@ public class ModelRendererWavefront
 
 	public void dispose()
 	{
-		list.delete();
+		if(list != null)
+			list.delete();
 	}
 
 	@Override
